Add return type for read() and write() method
