+++ universal_transfer_operator/src/universal_transfer_operator/data_providers/filesystem/local.py
@@ -10,7 +10,8 @@ import pandas as pd
 import smart_open
 from airflow.hooks.base import BaseHook
 
-from universal_transfer_operator.data_providers.filesystem.base import BaseFilesystemProviders, FileStream
+from universal_transfer_operator.data_providers.base import FileStream
+from universal_transfer_operator.data_providers.filesystem.base import BaseFilesystemProviders
 
 
 class LocalDataProvider(BaseFilesystemProviders):
@@ -64,7 +65,7 @@ class LocalDataProvider(BaseFilesystemProviders):
         """Return true if the dataset exists"""
         return exists(self.dataset.path)
 
-    def write_using_smart_open(self, source_ref: FileStream | pd.DataFrame):
+    def write_using_smart_open(self, source_ref: FileStream | pd.DataFrame) -> str:
         """Write the source data from remote object i/o buffer to the dataset using smart open"""
         mode = "wb" if self.read_as_binary(self.dataset.path) else "w"
         with smart_open.open(self.dataset.path, mode=mode, transport_params=self.transport_params) as stream:
